Require Clojure 1.5 so duplicate map keys do not throw exceptions.
JIRA ticket CLJ-1065, fixed in Clojure 1.5, changed the behaviour
of the hash-map constructor to allow duplicate keys, keeping the value
of the last one. Ideally a plist <dict> would not have duplicate keys,
but I've found them in iTune's Library XML, for example.

However, this means that now it's possible to lose data when parsing.
The alternative would be to allow passing in a merge function to be used
when the same key appears more than once in a dict.